地址管理优化

+++ foodie-pojo/src/main/java/com/foodie/entity/AddressBook.java
@@ -10,6 +10,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -18,7 +19,8 @@ import java.time.LocalDateTime;
  */
 @Data
 @TableName("address_book")
-public class AddressBook implements Serializable {
+public class
+AddressBook implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
@@ -68,10 +70,10 @@ public class AddressBook implements Serializable {
     private String label;
 
     /** 经度 */
-    private Double longitude;
+    private BigDecimal longitude;
 
     /** 纬度 */
-    private Double latitude;
+    private BigDecimal latitude;
 
     /**
      * 是否默认地址
